refactor: move v02 registration fns into v02 submodule
